Initial commit.
Not happy with the paths. Ideally there is just the landing page, which acts as
the long tweet editor if there is no query. Something very meaningful if you
type in the URL.

Pulled in jQuery reflexively. This is so simply, jQuery is not needed. But this
got done so quickly, I don't know that I care to remove jQuery and spar with
any unknown cross-browser issues.

Able to post the Gettysburg Address. Need to get the initial prompt to post
longer looking good, so that the uninitated user knows what do do.

Added bonus. Markdown!